Rename restraints_json to io
